docs: add root README, update MapScreen zoom argument type to double, commit IDE configs

+++ android-app/app/src/main/java/com/lixionary/pik2bus/ui/MapScreen.kt
@@ -82,7 +82,7 @@ fun MapScreen(
                             mapboxMap.moveCamera(
                                 CameraUpdateFactory.newLatLngZoom(
                                     MapboxLatLng(center.lat, center.lng),
-                                    (route.initial_zoom ?: 12).toFloat()
+                                    (route.initial_zoom ?: 12).toDouble()
                                 )
                             )
                         }
